+++ HomeWork5/model/LoadFrom.java
@@ -1,4 +1,4 @@
-package HomeWork3;
+package HomeWork5.model;
 
 import java.io.Serializable;
