+++ HomeWork5/model/Nimale.java
@@ -1,35 +1,34 @@
-package HomeWork3;
+package HomeWork5.model;
 
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Human implements Serializable, Comparable<Human> 
+public abstract class Nimale implements Serializable
 {
     private String name;
     private String sex;
     private int age;
-    private Human mather;
-    private Human father;
-    private List<Human> children;
-
-    public Human(String name, String sex, int age, Human mather, Human father) 
-    {      
+    private Nimale mather;
+    private Nimale father;
+    private List<Nimale> children;
+    public Nimale(String name, String sex, int age, Nimale mather, Nimale father) 
+    {
+              
         this.name = name;
         this.sex = sex;
         this.age = age;
         this.mather = mather;
         this.father = father;
-        children = new ArrayList<>();
-        FamilyFree.addFamilyFree(this);   
+        children = new ArrayList<>();         
     }
 
-    public Human(String name, String sex, int age) 
+    public Nimale(String name, String sex, int age) 
     {
         this(name, sex, age, null,null); // поправил конструктор!
     }
-    
-    public void addChild(Human child)
+
+    public void addChild(Nimale child)
     {
         if (!children.contains(child))
         {
@@ -59,7 +58,7 @@ public class Human implements Serializable, Comparable<Human>
     @Override
     public String toString() 
     {
-        String res = "Имя:" + name +  ",\t Пол: " + sex + ", Возраст: " + age;
+        String res = "Имя: " + name +  ",\t Пол: " + sex + ", Возраст: " + age;
         if (this.mather != null)
         {
             res += ", Мать: " + mather.name;
@@ -68,7 +67,7 @@ public class Human implements Serializable, Comparable<Human>
         
         if (this.father != null)
         {
-            res += ", Отец: " + father.name;
+            res += ", Отец: " + father.name + ",";
         }
         else res += ", Отец: неизвестен,";
 
@@ -83,12 +82,12 @@ public class Human implements Serializable, Comparable<Human>
         return name;
     }
 
-    public Human getFather() 
+    public Nimale getFather() 
     {
         return father;
     }
 
-    public Human getMather() 
+    public Nimale getMather() 
     {
         return mather;
     }
@@ -101,17 +100,6 @@ public class Human implements Serializable, Comparable<Human>
     public void setAge(int age) 
     {
         this.age = age;
-    }
-
-    public void saveObj(IO save)
-    {
-        save.save("FreeFamily.data", this);
-    }
-
-    @Override
-    public int compareTo(Human o) 
-    {
-        return name.compareToIgnoreCase(o.name);
     }
 
 }
